fix: 사용자 프로필 url 수정 및 사용자 확인 로직 업데이트
