+++ src/main/java/com/smooth/smooth_backend_user/user/controller/UserController.java
@@ -10,12 +10,17 @@ import com.smooth.smooth_backend_user.user.service.UserService;
 import com.smooth.smooth_backend_user.global.exception.BusinessException;
 import com.smooth.smooth_backend_user.user.exception.UserErrorCode;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.context.request.RequestContextHolder;
+import org.springframework.web.context.request.ServletRequestAttributes;
+import jakarta.servlet.http.HttpServletRequest;
 
+@Slf4j
 @RestController
 @RequestMapping("/api/users")
 @RequiredArgsConstructor
@@ -30,6 +35,21 @@ public class UserController {
             GatewayUserDetails userDetails = (GatewayUserDetails) authentication.getPrincipal();
             return userDetails.getUserId();
         }
+        
+        // Fallback: HTTP 헤더에서 직접 읽기
+        ServletRequestAttributes attr = (ServletRequestAttributes) RequestContextHolder.currentRequestAttributes();
+        HttpServletRequest request = attr.getRequest();
+        String userIdHeader = request.getHeader("X-User-Id");
+        String authenticatedHeader = request.getHeader("X-Authenticated");
+        
+        if (userIdHeader != null && "true".equals(authenticatedHeader)) {
+            try {
+                return Long.valueOf(userIdHeader);
+            } catch (NumberFormatException e) {
+                log.error("Invalid userId format in header: {}", userIdHeader);
+            }
+        }
+        
         throw new BusinessException(UserErrorCode.INVALID_TOKEN, "인증되지 않은 사용자입니다.");
     }
 
@@ -87,6 +107,17 @@ public class UserController {
 
         return ResponseEntity.ok(
                 ApiResponse.success("사용자 정보 조회 성공", info)
+        );
+    }
+
+    // 회원탈퇴
+    @DeleteMapping("/account")
+    public ResponseEntity<ApiResponse<Void>> deleteAccount() {
+        Long userId = getCurrentUserId();
+        userService.deleteAccount(userId);
+
+        return ResponseEntity.ok(
+                ApiResponse.success("회원탈퇴가 완료되었습니다.")
         );
     }
